requirements.txt
azure-functions
azure-identity
azure-ai-ml
